move aura consensus data provider to the node
