+++ client/rpc/src/eth/pending.rs
@@ -16,13 +16,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use std::{marker::PhantomData, sync::Arc};
-
 // Substrate
-use sc_client_api::{
-	backend::{AuxStore, Backend, StorageProvider},
-	UsageProvider,
-};
+use sc_client_api::backend::{Backend, StorageProvider};
 use sc_transaction_pool::ChainApi;
 use sc_transaction_pool_api::InPoolTransaction;
 use sp_api::{ApiExt, ApiRef, Core, ProvideRuntimeApi};
@@ -30,11 +25,10 @@ use sp_block_builder::BlockBuilder as BlockBuilderApi;
 use sp_blockchain::{ApplyExtrinsicFailed, HeaderBackend};
 use sp_inherents::{CreateInherentDataProviders, InherentData, InherentDataProvider};
 use sp_runtime::{
-	generic::{Digest, DigestItem},
+	generic::Digest,
 	traits::{Block as BlockT, Header as HeaderT, One},
 	TransactionOutcome,
 };
-use sp_timestamp::TimestampInherentData;
 
 use crate::eth::Eth;
 use fp_rpc::EthereumRuntimeRPCApi;
@@ -166,62 +160,5 @@ impl<B: BlockT> ConsensusDataProvider<B> for () {
 		_: &InherentData,
 	) -> Result<Digest, sp_inherents::Error> {
 		Ok(Default::default())
-	}
-}
-
-pub use self::aura::AuraConsensusDataProvider;
-mod aura {
-	use super::*;
-	use sp_consensus_aura::{
-		digests::CompatibleDigestItem,
-		sr25519::{AuthorityId, AuthoritySignature},
-		AuraApi, Slot, SlotDuration,
-	};
-
-	/// Consensus data provider for Aura.
-	pub struct AuraConsensusDataProvider<B, C> {
-		// slot duration
-		slot_duration: SlotDuration,
-		// phantom data for required generics
-		_phantom: PhantomData<(B, C)>,
-	}
-
-	impl<B, C> AuraConsensusDataProvider<B, C>
-	where
-		B: BlockT,
-		C: AuxStore + ProvideRuntimeApi<B> + UsageProvider<B>,
-		C::Api: AuraApi<B, AuthorityId>,
-	{
-		/// Creates a new instance of the [`AuraConsensusDataProvider`], requires that `client`
-		/// implements [`sp_consensus_aura::AuraApi`]
-		pub fn new(client: Arc<C>) -> Self {
-			let slot_duration = sc_consensus_aura::slot_duration(&*client)
-				.expect("slot_duration is always present; qed.");
-			Self {
-				slot_duration,
-				_phantom: PhantomData,
-			}
-		}
-	}
-
-	impl<B: BlockT, C: Send + Sync> ConsensusDataProvider<B> for AuraConsensusDataProvider<B, C> {
-		fn create_digest(
-			&self,
-			_parent: &B::Header,
-			data: &InherentData,
-		) -> Result<Digest, sp_inherents::Error> {
-			let timestamp = data
-				.timestamp_inherent_data()?
-				.expect("Timestamp is always present; qed");
-
-			let digest_item =
-				<DigestItem as CompatibleDigestItem<AuthoritySignature>>::aura_pre_digest(
-					Slot::from_timestamp(timestamp, self.slot_duration),
-				);
-
-			Ok(Digest {
-				logs: vec![digest_item],
-			})
-		}
 	}
 }
